requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
SQLAlchemy>=0.8.4,<=0.9.99,!=0.9.0,!=0.9.1,!=0.9.2,!=0.9.3,!=0.9.4,!=0.9.5,!=0.9.6
anyjson>=0.3.3,<=0.3.3
argparse
boto>=2.32.1,<2.35.0
decorator>=3.4.0,<=3.4.0
eventlet>=0.15.1,<=0.15.2
Jinja2<=2.7.2
keystonemiddleware>=1.0.0,<1.4.0
kombu>=2.5.0,<=3.0.7
lxml>=2.3,<=3.3.3
Routes>=1.12.3,!=2.0,<=2.1
WebOb>=1.2.3,<=1.3.1
greenlet>=0.3.2,<=0.4.2
PasteDeploy>=1.5.0,<=1.5.2
Paste<=1.7.5.1
sqlalchemy-migrate==0.9.1
netaddr>=0.7.12,<=0.7.13
suds==0.4
paramiko>=1.13.0,<=1.15.2
posix_ipc<=0.9.9
pyasn1<=0.1.7
Babel>=1.3,<=1.3
iso8601>=0.1.9,<=0.1.10
jsonschema>=2.0.0,<3.0.0
python-cinderclient>=1.1.0,<=1.1.1
python-neutronclient>=2.3.6,<2.4.0
python-glanceclient>=0.14.0,<=0.15.0
python-keystoneclient>=0.10.0,<1.2.0
six>=1.7.0,<=1.9.0
stevedore>=1.0.0,<=1.3.0 # Apache-2.0
websockify>=0.6.0,<0.7
wsgiref>=0.1.2
oslo.config>=1.4.0,<=1.6.0 # Apache-2.0
oslo.db>=1.0.0,<1.1  # Apache-2.0
oslo.rootwrap>=1.3.0,<=1.5.0
pycadf>=0.6.0,!=0.6.2,<0.7.0  # Apache-2.0
oslo.messaging>=1.4.0,<1.5.0
oslo.i18n>=1.0.0,<=1.3.1 # Apache-2.0
lockfile>=0.8,<=0.8
simplejson>=2.2.0,<=3.3.1
rfc3986>=0.2.0,<=0.2.0 # Apache-2.0
oslo.vmware>=0.6.0,<0.9.0               # Apache-2.0
python-dateutil>=2.4.0
psutil>=3.2.0
